# Ensure globbed files paths are expanded